+++ Cargo.toml
@@ -3,11 +3,15 @@ name = "nais_admin"
 version = "0.1.0"
 edition = "2018"
 
+[[bin]]
+name="nais_admin"
+path="src/main.rs"
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 candid = "0.7.4"
-clap = "2.33.3"
+clap = "3.0.0-beta.1"
 garcon = { version = "0.2.3", features = ["async"] }
 hex = "0.4.2"
 humantime = "2.0.1"
@@ -17,3 +21,6 @@ pem = "0.8.1"
 ring = "0.16.11"
 serde = "1.0.115"
 serde_json = "1.0.57"
+anyhow = "1.0.43"
+easy-hasher = "2.2.1"
+futures = "0.3"
