+++ src/main.rs
@@ -1,53 +1,36 @@
-use candid::{
-    parser::value::IDLValue,
-    types::{Function, Type},
-    CandidType, Decode, Deserialize, IDLArgs, IDLProg, TypeEnv,
-};
-use clap::{crate_authors, crate_version, AppSettings, Clap};
+pub mod command;
+
+use clap::{crate_authors, crate_version, Clap};
+use futures::executor::block_on;
 use ic_agent::{
-    agent::{self, signed::SignedUpdate, Replied},
-    agent::{
-        agent_error::HttpErrorPayload,
-        signed::{SignedQuery, SignedRequestStatus},
-    },
+    agent,
     export::Principal,
     identity::BasicIdentity,
-    Agent, AgentError, Identity,
+    Agent, Identity,
 };
 use ic_utils::interfaces::management_canister::{
-    builders::{CanisterInstall, CanisterSettings},
-    MgmtMethod,
+    builders::CanisterInstall,
 };
 use ring::signature::Ed25519KeyPair;
-use std::{
-    collections::VecDeque, convert::TryFrom, io::BufRead, path::PathBuf, process::exit,
-    str::FromStr,
-};
+use std::{fs::File, io::Read, path::PathBuf, str::FromStr};
 
 #[derive(Clap)]
 #[clap(
 version = crate_version!(),
-author = crate_authors!(),
-global_setting = AppSettings::GlobalVersion,
-global_setting = AppSettings::ColoredHelp
+author = crate_authors!()
 )]
-struct Opts {
-    /// Some input. Because this isn't an Option<T> it's required to be used
+pub struct Opts {
     #[clap(default_value = "http://localhost:8000/")]
     replica: String,
 
-    /// An optional PEM file to read the identity from. If none is passed,
-    /// a random identity will be created.
     #[clap(long)]
     pem: Option<PathBuf>,
 
-    /// An optional field to set the expiry time on requests. Can be a human
-    /// readable time (like `100s`) or a number of seconds.
     #[clap(long)]
     ttl: Option<humantime::Duration>,
 
     #[clap(subcommand)]
-    subcommand: SubCommand,
+    subcommand: command::SubCommand,
 }
 
 fn create_identity(maybe_pem: Option<PathBuf>) -> impl Identity {
@@ -69,16 +52,35 @@ fn create_identity(maybe_pem: Option<PathBuf>) -> impl Identity {
 fn main() {
     let opts: Opts = Opts::parse();
 
-    let my_identity = create_identity(opts.pem);
+    let my_identity = create_identity(opts.pem.clone());
     let my_principal =  my_identity.sender().expect("not valid identity");
 
     let agent = Agent::builder()
         .with_transport(
             agent::http_transport::ReqwestHttpReplicaV2Transport::create(opts.replica.clone())
-                .context("Failed to create Transport for Agent")?,
+                .expect("Failed to create Transport for Agent"),
         )
         .with_boxed_identity(Box::new(my_identity))
         .build()
-        .context("Failed to build the Agent")?;
+        .expect("Failed to build the Agent");
+
+    let mut f = File::open("nft.wasm").expect("no wasm file");
+    let mut buffer = Vec::<u8>::new();
+    f.read_to_end(&mut buffer).expect("read file error");
+    let code = easy_hasher::easy_hasher::Hash::from_vec(&buffer).to_hex_string();
+
+    let copts = command::CallOpts{
+        canister_id: Principal::from_text("rrkah-fqaaa-aaaaa-aaaaq-cai").unwrap(),
+        serialize: false,
+        candid: Some(PathBuf::from_str("nais_canister.did").unwrap()),
+        method_name: "uploadWasm".to_string(),
+        arg: command::ArgType::Idl,
+        output: command::ArgType::Idl,
+        arg_value: Some(code),
+    };
+
+    let future = command::call_update_method(&agent, &opts, &copts);
+    let result  = block_on(future);
+    command::show_result(result, &copts);
 
 }
